Address Greptile review: shared module, BANNER_KEY, file IO guard
- Extract humanize_seconds into Banner::Shared mixin (following
  Customers::Shared pattern) so set and show share one definition
- Add BANNER_KEY constant to Banner::Shared; all four CLI files
  reference it instead of the bare 'global_banner' string
- Guard File.read with rescue for Errno::EACCES/EISDIR so
  permission errors produce a clean message, not a backtrace
