+++ lib/onetime/cli.rb
@@ -113,6 +113,7 @@ require_relative 'cli/totp_command'
 require_relative 'cli/ratelimit_command'
 
 # Load banner CLI commands
+require_relative 'cli/banner/shared'
 require_relative 'cli/banner_command'
 require_relative 'cli/banner/set_command'
 require_relative 'cli/banner/show_command'
